agent/requirements.txt
fastapi
uvicorn
psycopg2-binary
structlog
pandas
requests
apscheduler
python-multipart
scikit-learn
numpy
prophet
